# docs/swagger.yaml
definitions:
  models.Employee:
    properties:
      email:
        type: string
      first_name:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
  models.ResponseData:
    properties:
      data: {}
      error: {}
      errors: {}
      message:
        type: string
    type: object
info:
  contact:
    email: mibnufaisal@gmail.com
    name: Muhammad Ibn Faisal
    url: https://github.com/ibnufaisal-27/mekari
  description: This is a repository of RESTFul API service.
  title: RESTFul API Service
  version: "1.0"
paths:
  /employees:
    delete:
      consumes:
      - application/json
      description: Delete Data Employee by ID
      operationId: delete-employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.Employee'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: Delete Employee data
    get:
      consumes:
      - application/json
      description: Retrive Data All Employee
      operationId: get-employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.Employee'
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: Get All Employee
    post:
      consumes:
      - application/json
      description: Create New Employee
      operationId: create-employee
      parameters:
      - description: Employee Data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.Employee'
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: Create New Employee based on payload
    put:
      consumes:
      - application/json
      description: Update Data Employee by ID
      operationId: update-employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.Employee'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: Update Employee data
  /employees/:id:
    get:
      consumes:
      - application/json
      description: Find Data Employee by ID
      operationId: find-employee
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/models.Employee'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseData'
            - properties:
                code:
                  type: integer
                message:
                  type: string
              type: object
      summary: Find Employee detail
swagger: "2.0"